GraphQL with FastAPI and Strawberry/requirements.txt
alembic==1.7.7
aniso8601==7.0.0
anyio==3.6.1
asgiref==3.5.2
backports.cached-property==1.0.1
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
promise==2.3
psycopg2==2.9.3
pydantic==1.9.1
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
strawberry-graphql==0.113.0
typing_extensions==4.2.0
uvicorn==0.17.6
websockets==10.3
watchgod==0.8.2
